requirements.txt
 
Django>=4.0
djangorestframework>=3.14.0
django-ckeditor>=6.5.1
django-redis>=5.2.0
googletrans==4.0.0-rc1
